# src/swagger.yml
openapi: 3.0.0
paths:
  /version:
    get:
      operationId: AppController_root
      summary: Service version
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInfo'
  /login:
    post:
      operationId: LoginController_login
      summary: get jwt
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /user/get/{id}:
    get:
      operationId: UsersController_get
      summary: Get user by id
      parameters:
        - name: id
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required: &ref_0
                  - statusCode
                  - message
          description: ''
      security:
        - bearer: []
        - token: []
  /user/register:
    post:
      operationId: UsersController_register
      summary: Register new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required: *ref_0
          description: ''
info:
  title: otus-highload
  description: ''
  version: 0.0.1
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: apiKey
      in: query
      name: token
  schemas:
    AppInfo:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
      required:
        - name
        - version
    LoginRequest:
      type: object
      properties:
        id:
          type: string
        password:
          type: string
      required:
        - id
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    User:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        second_name:
          type: string
        age:
          type: number
        birthdate:
          type: string
        biography:
          type: string
        city:
          type: string
      required:
        - id
        - first_name
        - second_name
        - age
        - birthdate
        - biography